add classes, refine code

+++ bin/password_manager.dart
@@ -1,7 +1,9 @@
+import 'package:password_manager/EncryptionManager.dart';
+import 'package:password_manager/HashingManager.dart';
+import 'package:password_manager/JsonManager.dart';
 import 'package:password_manager/password_maker.dart' as pass_gen;
 import 'package:encrypt/encrypt.dart' as encrypt;
 import 'package:convert/convert.dart';
-import 'package:argon2/argon2.dart' as argon2;
 import 'package:pointycastle/export.dart'
     show PBKDF2KeyDerivator, HMac, SHA256Digest, Pbkdf2Parameters;
 import 'dart:typed_data';
@@ -48,56 +50,61 @@ Future<void> manager() async {
         await findPassword(file);
         break;
       case '3':
-        await file.delete();
-        print(
-          'File deleted.\nYou will be able to create a new one next time you access the manager.',
-        );
+        print("ARE YOU SURE YOU WANT TO DELETE ALL YOUR DATA? (yes/no)");
+        final confirm = _getUserInput(['yes', 'no']);
+        if (confirm == 'yes') {
+          await file.delete();
+          print(
+            'File deleted.\nYou will be able to create a new one next time you access the manager.',
+          );
+        }
         exit(0);
     }
   } while (true);
 }
 
 Future<bool> accessControll() async {
+  JsonManager jsonManager = JsonManager(File('passwords.json'));
+  HashingManager hashingManager = HashingManager();
   try {
-    final file = File('passwords.json');
+    final File file = File('passwords.json');
 
     if (!await file.exists()) {
       print('First time using the manager. Creating a new file...');
       await file.create();
 
-      final masterPassword = _getNonEmptyInput('Insert the master password:');
-      final salt = generateSalt();
+      final String masterPassword =
+          _getNonEmptyInput('Insert the master password:');
+      final Uint8List salt = generateSalt();
 
       final data = {
         'passwords': [
           {
             'service': 'master',
-            'password': await passwordHashing(masterPassword, salt),
+            'password': await hashingManager.hashPassword(masterPassword, salt),
             'salt': hex.encode(salt),
           },
         ],
       };
 
-      await file.writeAsString(jsonEncode(data));
+      jsonManager.writeJsonFile(data);
       print('Master password saved successfully.');
       return true;
     } else {
-      final fileContent = await file.readAsString();
-
-      final masterPassword = _getNonEmptyInput(
+      final String masterPassword = _getNonEmptyInput(
         'Please insert the master password:',
       );
-      final data = jsonDecode(fileContent) as Map<String, dynamic>;
-      final saltHex = data['passwords'][0]['salt'];
-      final salt = Uint8List.fromList(hex.decode(saltHex));
+      final Map<String, dynamic> data = await jsonManager.readJsonFile();
+      final Uint8List salt = Uint8List.fromList(
+        hex.decode(data['passwords'][0]['salt']),
+      );
 
-      final hashedInput = await passwordHashing(masterPassword, salt);
-
-      for (final password in data['passwords']) {
-        if (password['service'] == 'master' &&
-            password['password'] == hashedInput) {
-          return true;
-        }
+      if (await hashingManager.verifyPassword(
+        masterPassword,
+        data['passwords'][0]['password'],
+        salt,
+      )) {
+        return true;
       }
       return false;
     }
@@ -111,16 +118,16 @@ Future<bool> accessControll() async {
 }
 
 Future<void> addNewPassword(File file) async {
-  final service = _getNonEmptyInput('Insert the service name:');
+  final String service = _getNonEmptyInput('Insert the service name:');
   print('Creating a password for $service...');
-  final password = pass_gen.passGen();
+  final String password = pass_gen.passGen();
   await encryptPassword(service, password, file);
 }
 
 Future<void> findPassword(File file) async {
   try {
     final service = _getNonEmptyInput('Insert the service name:');
-    final data = jsonDecode(await file.readAsString()) as Map<String, dynamic>;
+    final data = await JsonManager(file).readJsonFile();
     Uint8List? salt;
     var masterPassword = '';
 
@@ -132,16 +139,16 @@ Future<void> findPassword(File file) async {
 
       if (password['service'] == service && salt != null) {
         final key = deriveKey(masterPassword, salt, 32); // 32 bytes = 256 bits
-        final iv = encrypt.IV.fromBase64(password['iv']);
+        final iv = encrypt.IV.fromBase64(password[
+            'iv']); // Assicurati che l'IV venga decodificato correttamente da base64
 
-        final encrypter = encrypt.Encrypter(encrypt.AES(encrypt.Key(key)));
+        final decrypted = EncryptionManager().decryptData(
+          password['password'],
+          key,
+          iv,
+        );
+        print('Password found: $decrypted');
 
-        try {
-          final decrypted = encrypter.decrypt64(password['password'], iv: iv);
-          print('Password found: $decrypted');
-        } catch (e) {
-          print('Error during decryption: $e');
-        }
         return;
       }
     }
@@ -152,31 +159,6 @@ Future<void> findPassword(File file) async {
     print('File read error: $e');
   } catch (e) {
     print('An unexpected error occurred: $e');
-  }
-}
-
-Future<String> passwordHashing(String password, Uint8List salt) async {
-  try {
-    final parameters = argon2.Argon2Parameters(
-      argon2.Argon2Parameters.ARGON2_id,
-      salt,
-      version: argon2.Argon2Parameters.ARGON2_VERSION_13,
-      iterations: 6,
-      memoryPowerOf2: 16,
-    );
-
-    final argon2Generator = argon2.Argon2BytesGenerator();
-
-    argon2Generator.init(parameters);
-
-    final passwordBytes = parameters.converter.convert(password);
-    final result = Uint8List(32);
-    argon2Generator.generateBytes(passwordBytes, result, 0, result.length);
-
-    return hex.encode(result);
-  } catch (e) {
-    print('Error while hashing the password: $e');
-    throw Exception('Hashing failed');
   }
 }
 
@@ -197,16 +179,18 @@ Future<void> encryptPassword(String service, String password, File file) async {
         masterPasswordFound = true;
         final salt = Uint8List.fromList(hex.decode(passwordEntry['salt']));
         final key = deriveKey(passwordEntry['password'], salt, 32);
-        final iv = encrypt.IV.fromLength(16);
-        final encrypter = encrypt.Encrypter(encrypt.AES(encrypt.Key(key)));
-        final encrypted = encrypter.encrypt(password, iv: iv);
+
+        // Aggiungi la generazione dell'IV (Initialization Vector)
+        final iv = encrypt.IV.fromLength(16); // Genera un IV casuale di 16 byte
+        final encrypted = EncryptionManager().encryptData(password, key, iv);
 
         print('Password encrypted successfully.');
 
         newPasswords.add({
           'service': service,
-          'password': encrypted.base64,
-          'iv': iv.base64,
+          'password': encrypted,
+          'iv': iv
+              .base64, // Assicurati che l'IV venga salvato correttamente in base64
         });
       }
     }
@@ -215,8 +199,9 @@ Future<void> encryptPassword(String service, String password, File file) async {
       throw Exception('Master password not found.');
     }
 
-    data['passwords'].addAll(newPasswords);
-    await file.writeAsString(jsonEncode(data));
+    for (var passwordData in newPasswords) {
+      await JsonManager(file).updateJsonFile(passwordData);
+    }
     print('$service password saved successfully.');
   } catch (e) {
     print('Error encrypting password: $e');
